Terminal Api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pymssql==2.3.0
pydantic==2.7.3
pytest==8.2.1
httpx==0.27.0
black==24.4.2
isort==5.13.2
ruff==0.4.8
